+++ src/fields/MultiSelect.php
@@ -10,6 +10,7 @@ namespace craft\fields;
 use Craft;
 use craft\base\ElementInterface;
 use craft\fields\data\MultiOptionsFieldData;
+use craft\helpers\ArrayHelper;
 
 /**
  * MultiSelect represents a Multi-select field.
@@ -50,6 +51,11 @@ class MultiSelect extends BaseOptionsField
      */
     protected function inputHtml($value, ElementInterface $element = null): string
     {
+        /** @var MultiOptionsFieldData $value */
+        if (ArrayHelper::contains($value, 'valid', false, true)) {
+            Craft::$app->getView()->setInitialDeltaValue($this->handle, null);
+        }
+
         return Craft::$app->getView()->renderTemplate('_includes/forms/multiselect', [
             'name' => $this->handle,
             'values' => $value,