+++ src/fields/RadioButtons.php
@@ -41,6 +41,11 @@ class RadioButtons extends BaseOptionsField implements SortableFieldInterface
      */
     protected function inputHtml($value, ElementInterface $element = null): string
     {
+        /** @var SingleOptionFieldData $value */
+        if (!$value->valid) {
+            Craft::$app->getView()->setInitialDeltaValue($this->handle, null);
+        }
+
         return Craft::$app->getView()->renderTemplate('_includes/forms/radioGroup', [
             'name' => $this->handle,
             'value' => $value,